A little bit of PMM work

+++ arch/x86_64/lib/inttypes.h
@@ -44,37 +44,37 @@
 #define SIZE_MAX        18446744073709551615
 
 typedef signed char         int8_t;
-typedef signed short        int16_t;
+typedef signed short int    int16_t;
 typedef signed int          int32_t;
-typedef signed long long    int64_t;
-typedef signed long long    intmax_t;
-typedef signed long long    intptr_t;
+typedef signed long int     int64_t;
+typedef signed long int     intmax_t;
+typedef signed long int     intptr_t;
 
 typedef signed int          int_least8_t;
 typedef signed int          int_least16_t;
 typedef signed int          int_least32_t;
-typedef signed long long    int_least64_t;
+typedef signed long int     int_least64_t;
 
 typedef signed int          int_fast8_t;
 typedef signed int          int_fast16_t;
 typedef signed int          int_fast32_t;
-typedef signed long long    int_fast64_t;
+typedef signed long int     int_fast64_t;
 
 typedef unsigned char       uint8_t;
-typedef unsigned short      uint16_t;
+typedef unsigned short int  uint16_t;
 typedef unsigned int        uint32_t;
-typedef unsigned long long  uint64_t;
-typedef unsigned long long  uintmax_t;
-typedef unsigned long long  uintptr_t;
+typedef unsigned long int   uint64_t;
+typedef unsigned long int   uintmax_t;
+typedef unsigned long int   uintptr_t;
 
 typedef unsigned int        uint_least8_t;
 typedef unsigned int        uint_least16_t;
 typedef unsigned int        uint_least32_t;
-typedef unsigned long long  uint_least64_t;
+typedef unsigned long int   uint_least64_t;
 
 typedef unsigned int        uint_fast8_t;
 typedef unsigned int        uint_fast16_t;
 typedef unsigned int        uint_fast32_t;
-typedef unsigned long long  uint_fast64_t;
+typedef unsigned long int   uint_fast64_t;
 
 #endif
